--- llm_council/members/alibaba.yaml
@@ -1,69 +1,0 @@
-- model_name: Qwen2-72B-Instruct
-  providers:
-    - name: together
-      fully_qualified_name: together://Qwen/Qwen2-72B-Instruct
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.9
-          per_1m_output_tokens: 0.9
-    - name: lepton
-      fully_qualified_name: lepton://qwen2-72b
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.8
-          per_1m_output_tokens: 0.8
-  model_info:
-    context_window: 32768
-    training_data_cutoff: "Unknown"
-    release_date: "August 3, 2023"
-    num_parameters: "72b"
-    modalities:
-      - text
-
-- model_name: Qwen1.5-72B-Chat
-  providers:
-    - name: together
-      fully_qualified_name: together://Qwen/Qwen1.5-72B-Chat
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.9
-          per_1m_output_tokens: 0.9
-  model_info:
-    context_window: 32768
-    training_data_cutoff: "Unknown"
-    release_date: "February 4, 2024"
-    num_parameters: 72b
-    modalities:
-      - text
-
-- model_name: Qwen1.5-32B-Chat
-  providers:
-    - name: together
-      fully_qualified_name: together://Qwen/Qwen1.5-32B-Chat
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.8
-          per_1m_output_tokens: 0.8
-  model_info:
-    context_window: 32768
-    training_data_cutoff: "Unknown"
-    release_date: "February 4, 2024"
-    num_parameters: "32b"
-    modalities:
-      - text
-
-- model_name: Qwen1.5-110B-Chat
-  providers:
-    - name: together
-      fully_qualified_name: together://Qwen/Qwen1.5-110B-Chat
-      costs:
-        serverless:
-          per_1m_input_tokens: 1.8
-          per_1m_output_tokens: 1.8
-  model_info:
-    context_window: 32768
-    training_data_cutoff: "Unknown"
-    release_date: "February 4, 2024"
-    num_parameters: "110b"
-    modalities:
-      - text
