--- llm_council/members/anthropic.yaml
@@ -1,82 +1,0 @@
-# Schema defined in llm_council/members/schema.py
-- model_name: claude-3-haiku
-  providers:
-    - name: anthropic
-      fully_qualified_name: anthropic://claude-3-haiku-20240307
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.25
-          per_1m_output_tokens: 1.25
-  model_info:
-    context_window: 200000
-    training_data_cutoff: "Aug 2023"
-    release_date: "March 7, 2024"
-    num_parameters: "Unknown"
-    modalities:
-      - text
-      - image
-
-- model_name: claude-3-sonnet
-  providers:
-    - name: anthropic
-      fully_qualified_name: anthropic://claude-3-sonnet-20240229
-      costs:
-        serverless:
-          per_1m_input_tokens: 3.00
-          per_1m_output_tokens: 15.00
-  model_info:
-    context_window: 200000
-    training_data_cutoff: "Aug 2023"
-    release_date: "Feb 29, 2024"
-    num_parameters: "Unknown"
-    modalities:
-      - text
-  status: obsolescent
-
-- model_name: claude-3-opus
-  providers:
-    - name: anthropic
-      fully_qualified_name: anthropic://claude-3-opus-20240229
-      costs:
-        serverless:
-          per_1m_input_tokens: 15.00
-          per_1m_output_tokens: 75.00
-  model_info:
-    context_window: 200000
-    training_data_cutoff: "Aug 2023"
-    release_date: "Feb 29, 2024"
-    num_parameters: "Unknown"
-    modalities:
-      - text
-      - image
-
-- model_name: claude-3-5-sonnet
-  providers:
-    - name: anthropic
-      fully_qualified_name: anthropic://claude-3-5-sonnet-20240620
-      status: obsolescent
-      costs:
-        serverless:
-          per_1m_input_tokens: 3.00
-          per_1m_output_tokens: 15.00
-    - name: anthropic
-      fully_qualified_name: anthropic://claude-3-5-sonnet-20241022
-      costs:
-        serverless:
-          per_1m_input_tokens: 3.00
-          per_1m_output_tokens: 15.00
-    - name: anthropic
-      fully_qualified_name: anthropic://claude-3-5-sonnet-latest
-      costs:
-        serverless:
-          per_1m_input_tokens: 3.00
-          per_1m_output_tokens: 15.00
-  model_info:
-    context_window: 200000
-    max_output_tokens: "8192"
-    training_data_cutoff: "Apr 2024"
-    release_date: "June 20, 2024"
-    num_parameters: "Unknown"
-    modalities:
-      - text
-      - image
