--- llm_council/members/cohere.yaml
@@ -1,62 +1,0 @@
-# https://docs.cohere.com/v2/docs/command-r-plus
-- model_name: command-r-plus
-  providers:
-    - name: cohere
-      fully_qualified_name: cohere://command-r-plus-08-2024
-      max_output_tokens: 4k
-      costs:
-        serverless:
-          per_1m_input_tokens: 2.50
-          per_1m_output_tokens: 10
-    - name: cohere
-      fully_qualified_name: cohere://command-r-plus-04-2024
-      max_output_tokens: 4k
-      costs:
-        serverless:
-          per_1m_input_tokens: 2.50
-          per_1m_output_tokens: 10
-    - name: cohere
-      fully_qualified_name: cohere://command-r-plus
-      max_output_tokens: 4k
-      costs:
-        serverless:
-          per_1m_input_tokens: 2.50
-          per_1m_output_tokens: 10
-  model_info:
-    context_window: 128k
-    training_data_cutoff: Unknown
-    release_date: Unknown
-    num_parameters: Unknown
-    modalities:
-      - text
-
-- model_name: command-r-plus
-  providers:
-    - name: cohere
-      fully_qualified_name: cohere://command-r-08-2024
-      max_output_tokens: 4k
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.15
-          per_1m_output_tokens: 0.6
-    - name: cohere
-      fully_qualified_name: cohere://command-r-04-2024
-      max_output_tokens: 4k
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.15
-          per_1m_output_tokens: 0.6
-    - name: cohere
-      fully_qualified_name: cohere://command-r
-      max_output_tokens: 4k
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.15
-          per_1m_output_tokens: 0.6
-  model_info:
-    context_window: 128k
-    training_data_cutoff: Unknown
-    release_date: Unknown
-    num_parameters: Unknown
-    modalities:
-      - text
