--- llm_council/members/databricks.yaml
@@ -1,16 +1,0 @@
-- model_name: dbrx-instruct
-  providers:
-    - name: together
-      fully_qualified_name: together://databricks/dbrx-instruct
-      costs:
-        serverless:
-          per_1m_input_tokens: 1.2
-          per_1m_output_tokens: 1.2
-  model_info:
-    context_window: 32768
-    max_output_tokens: "0"
-    training_data_cutoff: "December 2023"
-    release_date: "March 27, 2024"
-    num_parameters: "132b"
-    modalities:
-      - text
