--- llm_council/members/google.yaml
@@ -1,61 +1,0 @@
-- model_name: gemini-1.0-pro
-  providers:
-    - name: vertex
-      fully_qualified_name: vertex://gemini-1.0-pro
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.125
-          per_1m_output_tokens: 0.375
-  model_info:
-    context_window: 32K
-    training_data_cutoff: Unknown
-    release_date: Unknown
-    num_parameters: Unknown
-    modalities:
-      - text
-      - image
-      - audio
-      - video
-
-- model_name: gemini-1.5-pro
-  providers:
-    - name: vertex
-      fully_qualified_name: vertex://gemini-1.5-pro-002
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.3125
-          per_1m_output_tokens: 1.25
-  model_info:
-    context_window: 128k
-    training_data_cutoff: Unknown
-    release_date: Unknown
-    num_parameters: Unknown
-    modalities:
-      - text
-      - image
-      - audio
-      - video
-
-- model_name: gemini-1.5-flash
-  providers:
-    - name: vertex
-      fully_qualified_name: vertex://gemini-1.5-flash-002
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.01875
-          per_1m_output_tokens: 0.075
-  model_info:
-    context_window: 128k
-    training_data_cutoff: Unknown
-    release_date: Unknown
-    num_parameters: Unknown
-    modalities:
-      - text
-      - image
-      - audio
-      - video
-
-# Vertex.
-# https://cloud.google.com/vertex-ai/generative-ai/pricing
-# Google.
-# https://ai.google.dev/pricing
