--- llm_council/members/mistral.yaml
@@ -1,49 +1,0 @@
-# Schema defined in llm_council/members/schema.py
-- model_name: mistral-large
-  providers:
-    - name: mistral
-      fully_qualified_name: mistral://mistral-large-latest
-      costs:
-        serverless:
-          per_1m_input_tokens: 2
-          per_1m_output_tokens: 6
-  short_name: mistral-large
-  model_info:
-    context_window: 128k
-    training_data_cutoff: "Unknown"
-    release_date: "February 26, 2024"
-    num_parameters: "Unknown"
-    modalities:
-      - text
-
-- model_name: mixtral-8x22b
-  providers:
-    - name: mistral
-      fully_qualified_name: mistral://open-mixtral-8x22b
-      costs:
-        serverless:
-          per_1m_input_tokens: 2
-          per_1m_output_tokens: 6
-  model_info:
-    context_window: 64k
-    training_data_cutoff: "Unknown"
-    release_date: "April 17, 2024"
-    num_parameters: "141b"
-    modalities:
-      - text
-
-- model_name: mixtral-8x7b
-  providers:
-    - name: mistral
-      fully_qualified_name: mistral://open-mixtral-8x7b
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.7
-          per_1m_output_tokens: 0.7
-  model_info:
-    context_window: 32k
-    training_data_cutoff: "Unknown"
-    release_date: "Dec 11, 2023"
-    num_parameters: "45B"
-    modalities:
-      - text
