--- llm_council/members/openai.yaml
@@ -1,153 +1,0 @@
-# Schema defined in llm_council/members/schema.py
-- model_name: gpt-4-turbo
-  providers:
-    - name: openai
-      fully_qualified_name: openai://gpt-4-turbo-2024-04-09
-      max_output_tokens: 4096
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.15
-          per_1m_input_tokens_cached: 0.075
-          per_1m_output_tokens: 0.6
-        batch:
-          per_1m_input_tokens: 0.075
-          per_1m_output_tokens: 0.3
-      status: obsolescent
-  model_info:
-    context_window: 128000
-    training_data_cutoff: Dec 2023
-    release_date: April 9, 2024
-    num_parameters: Unknown
-    modalities:
-      - text
-      - image
-
-- model_name: gpt-4
-  providers:
-    - name: openai
-      fully_qualified_name: openai://gpt-4-0613
-      costs:
-        serverless:
-          per_1m_input_tokens: 30
-          per_1m_output_tokens: 60
-  model_info:
-    context_window: 8192
-    max_output_tokens: 8192
-    training_data_cutoff: Sep 2021
-    release_date: June 13, 2023
-    num_parameters: Unknown
-    modalities:
-      - text
-  status: obsolescent
-
-- model_name: gpt-4o-mini
-  providers:
-    - name: openai
-      fully_qualified_name: openai://gpt-4o-mini-2024-07-18
-      max_output_tokens: 16384
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.15
-          per_1m_input_tokens_cached: 0.075
-          per_1m_output_tokens: 0.6
-        batch:
-          per_1m_input_tokens: 0.075
-          per_1m_output_tokens: 0.3
-    - name: openai
-      fully_qualified_name: openai://gpt-4o-mini
-      max_output_tokens: 16384
-      costs:
-        serverless:
-          per_1m_input_tokens: 0.15
-          per_1m_input_tokens_cached: 0.075
-          per_1m_output_tokens: 0.6
-        batch:
-          per_1m_input_tokens: 0.075
-          per_1m_output_tokens: 0.3
-  model_info:
-    context_window: 128000
-    training_data_cutoff: Up to Oct 2023
-    release_date: July 18, 2024
-    num_parameters: Unknown
-    modalities:
-      - text
-      - image
-
-- model_name: gpt-4o
-  providers:
-    - name: openai
-      fully_qualified_name: openai://gpt-4o-2024-08-06
-      max_output_tokens: 16384
-      costs:
-        serverless:
-          per_1m_input_tokens: 2.5
-          per_1m_input_tokens_cached: 1.25
-          per_1m_output_tokens: 10
-        batch:
-          per_1m_input_tokens: 1.25
-          per_1m_output_tokens: 5
-    - name: openai
-      fully_qualified_name: openai://gpt-4o-2024-05-13
-      max_output_tokens: 16384
-      costs:
-        serverless:
-          per_1m_input_tokens: 5
-          per_1m_output_tokens: 15
-        batch:
-          per_1m_input_tokens: 2.5
-          per_1m_output_tokens: 7.5
-    - name: openai
-      fully_qualified_name: openai://chatgpt-4o-latest
-      max_output_tokens: 16384
-      costs:
-        serverless:
-          per_1m_input_tokens: 2.5
-          per_1m_input_tokens_cached: 1.25
-          per_1m_output_tokens: 10
-        batch:
-          per_1m_input_tokens: 1.25
-          per_1m_output_tokens: 5
-  model_info:
-    context_window: 128000
-    training_data_cutoff: Up to Oct 2023
-    release_date: August 6, 2024
-    num_parameters: Unknown
-    modalities:
-      - text
-      - image
-
-- model_name: o1-preview
-  providers:
-    - name: openai
-      fully_qualified_name: openai://o1-preview-2024-09-12
-      max_output_tokens: 32768
-      costs:
-        serverless:
-          per_1m_input_tokens: 15
-          per_1m_input_tokens_cached: 7.5
-          per_1m_output_tokens: 60
-  model_info:
-    context_window: 128000
-    training_data_cutoff: Up to Oct 2023
-    release_date: September 12, 2024
-    num_parameters: Unknown
-    modalities:
-      - text
-
-- model_name: o1-mini
-  providers:
-    - name: openai
-      fully_qualified_name: openai://o1-mini-2024-09-12
-      max_output_tokens: 32768
-      costs:
-        serverless:
-          per_1m_input_tokens: 3
-          per_1m_input_tokens_cached: 1.5
-          per_1m_output_tokens: 12
-  model_info:
-    context_window: 128000
-    training_data_cutoff: Oct 2023
-    release_date: September 12, 2024
-    num_parameters: Unknown
-    modalities:
-      - text
