--- tests/llm_council/providers/test_providers.py
@@ -1,9 +1,0 @@
-import os
-import pytest
-
-from llm_council.providers.base_provider import PROVIDER_REGISTRY
-
-
-@pytest.mark.parametrize("service", PROVIDER_REGISTRY)
-def test_api_keys_exist(service):
-    assert PROVIDER_REGISTRY[service]["enabled"]
